backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
elasticsearch[async]==8.17.0
httpx==0.28.1
apscheduler==3.10.4
pydantic==2.10.4
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
